Always-on VPN support for API24+ devices
Starting with API24, Android supports an Always-On vpn mode. This can be turned-on by the user,
and in response OS will startup configured VPN service on device boot, and after app upgrades.
Essentially, Android will attempt to always keep the VPN service around.

This patch introduces basic support for this mode. Our service startup changes a little bit.
When it's the OS that's starting us (either during boot, app upgrade, or when user turns-on Always-on mode
in system settings), it will start us without any extras in the intent. When we detect this, kick-off
a full startup as it would happen if the user switched-on VPN from within the UI itself.

This does introduce a bit of duplication of functionality - but it's mainly around how "use cases" are
wired together. Reducing this duplication would ideally require a refactor of how we manage use cases,
and that's not something I want to do right now. Besides, this duplication is pretty trivial.

"Normal" service startup (i.e. initialized from within the app itself) is unchanged.